Fixed styling error in Users
