+++ api/apiRouter.js
@@ -1,12 +1,5 @@
-const express = require('express')
 const users = require('./routes/users')
 
-const api = express.Router()
-
-api.get('/', (req, res) => {
-    res.json("App is currently functioning")
-})
-
-api.use('/api', users)
-
-module.exports = api
+module.exports = (server) => {
+    server.use('/api', users)
+}
