--- api/helpers/codes.js
@@ -1,9 +1,0 @@
-
-module.exports = {
-    SUCCESS: {
-        code: 200,
-        message: "Success"
-    },
-    
-
-}
