+++ package.json
@@ -10,11 +10,13 @@
   "license": "MIT",
   "dependencies": {
     "bcrypt": "^3.0.0",
+    "cors": "^2.8.4",
     "express": "^4.16.3",
     "helmet": "^3.13.0",
     "jsonwebtoken": "^8.3.0",
     "knex": "^0.15.2",
     "morgan": "^1.9.0",
+    "react-router-dom": "^4.3.1",
     "sqlite3": "^4.0.2"
   },
   "devDependencies": {
