+++ server.js
@@ -1,15 +1,6 @@
-const express = require('express')
-const helmet = require('helmet')
-const morgan = require('morgan')
-const apiRoutes = require('./api/apiRouter')
-
-const server = express()
-server.use(helmet())
-server.use(morgan('tiny'))
-server.use(express.json())
-
-server.use('/', apiRoutes)
-
+const server = require('express')()
+require('./api/middleware')(server)
+require('./api/apiRouter')(server)
 
 const port = 8000
 server.listen(port, () => {
